:truck: Moved tests to `/tests` folder
