+++ packages/react-magic-motion/tests/magic-exit.test.tsx
