+++ packages/react-magic-motion/tests/magic-motion.test.tsx
@@ -1,50 +1,11 @@
 import { render } from "@testing-library/react";
 import { beforeAll, describe, expect, test, vi, afterEach } from "vitest";
 import "@testing-library/jest-dom";
-import { useRef, type ReactNode, forwardRef } from "react";
 import { motion } from "framer-motion";
 import { convertChildrenToMotionChildren } from "../utils/magic-animation";
 import { MagicExit } from "../magic-exit";
-import { MagicMotion } from ".";
-
-function TestComponent({
-  customText,
-  testId,
-}: {
-  customText: ReactNode;
-  testId: string;
-}): JSX.Element {
-  return <div data-testid={testId}>{customText}</div>;
-}
-
-function ParentComponent(): JSX.Element {
-  return <TestComponent customText="test" testId="string-child" />;
-}
-
-interface ForwardedRefComponent {
-  children?: React.ReactNode;
-}
-
-const ForwardedRefComponent = forwardRef<
-  HTMLButtonElement,
-  ForwardedRefComponent
->((props, ref) => {
-  return (
-    <button type="button" ref={ref} className="FancyButton">
-      {props.children}
-    </button>
-  );
-});
-
-ForwardedRefComponent.displayName = "ForwardedRefComponent";
-
-function ForwardedRefParent(): JSX.Element {
-  const buttonRef = useRef<HTMLButtonElement>(null); // Specify the element type
-
-  return (
-    <ForwardedRefComponent ref={buttonRef}>Click me!</ForwardedRefComponent>
-  );
-}
+import { MagicMotion } from "../magic-motion";
+import { TestComponent, ParentComponent, ForwardedRefParent } from ".";
 
 describe("<MagicMotion> tests", () => {
   const consoleMock = vi
